compartmentalized printing of output in cpu version

+++ include/cpu/constants.h
@@ -14,8 +14,8 @@
 
 typedef int count_t;
 
-typedef double num_t;
-#define NUM_TF "%f"
+typedef float num_t;
+#define NUM_TF "%0.3f"
 
 
 #endif /* CONSTANTS_H */
